+++ src/state/entity/ig_ENTITY_Magnet.ts
@@ -1,0 +1,38 @@
+import { registerNetEntity } from '../../misc/entity-netid'
+import { prestart } from '../../loading-stages'
+import { StateMemory } from '../state-util'
+import type { StateKey } from '../states'
+import * as igAnimatedEntity from './ig_AnimatedEntity-base'
+
+declare global {
+    namespace ig.ENTITY {
+        interface Magnet extends StateMemory.MapHolder<StateKey> {}
+    }
+    interface EntityStates {
+        'ig.ENTITY.Magnet': Return
+    }
+}
+
+type Return = ReturnType<typeof getEntityState>
+function getEntityState(this: ig.ENTITY.Magnet, player?: StateKey) {
+    const memory = StateMemory.getBy(this, player)
+
+    return {
+        ...igAnimatedEntity.getEntityState.call(this, memory),
+    }
+}
+
+function setEntityState(this: ig.ENTITY.Magnet, state: Return) {
+    igAnimatedEntity.setEntityState.call(this, state)
+}
+
+prestart(() => {
+    ig.ENTITY.Magnet.inject({
+        getEntityState,
+        setEntityState,
+    })
+    ig.ENTITY.Magnet.create = () => {
+        throw new Error('ig.ENTITY.Magnet.create not implemented')
+    }
+    registerNetEntity({ entityClass: ig.ENTITY.Magnet, isStatic: true })
+}, 2)